[#19] refactor: :hammer: 리팩토링, 부족한 부분 추가 구현 - 파일 및 폴더 구조 변경    - common 폴더 생성 - Context.jsx    - useContext 사용하여 prop drilling 개선 - 팝업 이벤트 구현    - 수량 변경    - 주문하기    - 주문결과 안내 메시지 UI
