+++ src/common/api.js
@@ -1,6 +1,6 @@
-const api = () => {
+const API = () => {
     const URL = "https://codesquad-2021-api.herokuapp.com/sidedish";
-    const api = (path="", data="") => {
+    const API = path => {
         const req = {
             method : 'GET',
             headers: {
@@ -9,12 +9,12 @@ const api = () => {
         };
         return fetch(URL + path, req);
     }
-    return async(path, data) => {
-        const result = await api(path, data);
+    return async (path) => {
+        const result = await API(path);
         let json = await result.json();
         if (!json || json.length === 0) json = null;
         return json;
     }
 }
 
-export default api();
+export default API();
