+++ src/common/const.js
@@ -1,0 +1,15 @@
+const TEXT = {
+  "bestTab-title": "후기가 증명하는 베스트 반찬",
+  "order" : {
+    true : {
+      "title": "주문 성공",
+      "content": "감사합니다! \r\n맛있게 준비해드릴게요!"
+    },
+    false : {
+      "title": "주문 실패",
+      "content": "재고가 부족합니다!"
+    }
+  }
+};
+
+export { TEXT };
