+++ src/component/main/slideContainer/SlideContainer.jsx
