--- src/const.js
@@ -1,5 +1,0 @@
-const TEXT = {
-  "bestTab-title": "후기가 증명하는 베스트 반찬",
-};
-
-export { TEXT };
